use set instead of list
Lookups are O(n) in lists and O(1) in dictionaries and sets. I'll use a set because we don't need to associate values.
To count "(i, j)" and "(j, i)" once, we'll look up "j" in a copy of the set "numbers" (instead of the "numbers" itself), and then we'll delete "i" from the copy. This way when "i:=j", we won't find "j:=i" in the copy. Note that we can't modify the set while we're iterating over it, and also the numbers have to reset for each time we call "func(x)" (these are other reasons I use a copy of "numbers").
This solution is merely convenient and neither memory nor time-efficient.